requirements.txt
asn1crypto
certifi
cryptography
dblpy
discord.py
entrypoints
GitPython
grequests
keyring
keyrings.alt
pipdeptree
pipreqs
ptv-python-wrapper
pycrypto
PyNaCl
python-dateutil
pyxdg
setuptools
ssh-import-id
urllib3
wheel
psycopg2-binary
